+++ app/javascript/controllers/maps/maplibre_controller.js
@@ -93,7 +93,7 @@ export default class extends Controller {
 
     // Initialize managers
     this.layerManager = new LayerManager(this.map, this.settings, this.api)
-    this.dataLoader = new DataLoader(this.api, this.apiKeyValue)
+    this.dataLoader = new DataLoader(this.api, this.apiKeyValue, this.settings)
     this.eventHandlers = new EventHandlers(this.map, this)
     this.filterManager = new FilterManager(this.dataLoader)
     this.mapDataManager = new MapDataManager(this)